requirements.txt
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
asyncpg
attrs
backoff
beautifulsoup4
certifi
cffi
charset-normalizer
click
cryptography
dataclasses-json
emoji
faiss-cpu
fastapi-slim
filetype
frozenlist
greenlet
grpcio
h11
h2
hpack
html5lib
httpcore
httpx
httpx-sse
hyperframe
idna
joblib
jsonpatch
jsonpointer
langchain
langchain-community
langchain-core
langchain-ollama
langchain-qdrant
langchain-text-splitters
langchainhub
langdetect
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
lxml
marshmallow
multidict
mypy_extensions
nest-asyncio
nltk
numpy
olefile
ollama
openapi-pydantic
orjson
ormsgpack
packaging
portalocker
propcache
protobuf
psutil
psycopg2-binary
pycparser
pydantic
pydantic-settings
pydantic_core
pypdf
python-docx
python-dotenv
python-iso639
python-magic
python-multipart
python-oxmsg
PyYAML
qdrant-client
RapidFuzz
regex
requests
requests-toolbelt
six
sniffio
soupsieve
SQLAlchemy
starlette
tenacity
tiktoken
tqdm
types-requests
typing-inspect
typing-inspection
typing_extensions
unstructured
unstructured-client
urllib3
uvicorn
webencodings
wrapt
xxhash
yarl
zstandard
